feat(multiphase): add phasecount as envoy property
Collect CT Meter configuration from ivp/meters and determine number of
phases in use. Set added SupportedFeatures DUALPHASE or THREEPHASE
when phaseCount > 1 and CT's are enabled.

Add new property phase_count to Envoy that will return 1,2 or 3 based on
added properties.

Phase_count >1 or added properties to be used to handle multi phase
data for production and/or consumption.